+++ app/src/main/java/com/Skyflow/PullFunds.kt
@@ -126,8 +126,8 @@ class PullFunds : AppCompatActivity() {
             requestHeader.put("Accept", "application/json")
             requestHeader.put("Authorization",BuildConfig.GATEWAY_TOKEN)
             val url = BuildConfig.GATEWAY_URL_PULL_FUNDS
-            val bodyForGateway = GatewayConfiguration(gatewayURL = url,requestHeader = requestHeader,pathParams = pathparams,methodName = RequestMethod.POST, requestBody=reqBodyObj, responseBody =  responseBody,queryParams = queryParams)
-            skyflowClient.invokeGateway(bodyForGateway,object : Callback
+            val bodyForGateway = ConnectionConfiguration(connectionURL = url,requestHeader = requestHeader,pathParams = pathparams,methodName = RequestMethod.POST, requestBody=reqBodyObj, responseBody =  responseBody,queryParams = queryParams)
+            skyflowClient.invokeConnection(bodyForGateway,object : Callback
             {
                 override fun onSuccess(responseBody: Any) {
                     runOnUiThread(Runnable {